feat: The "session request" is no longer managed by the client. The hosting application is expected to manage the session request. This has presented a cleaner and more reliable API for us in production.
BREAKING CHANGE: The sessionRequest argument is no longer part of the constructor. The sessionRequested argument is no longe provided to the getRealtimeEphemeralAPIKey callback. The host application should know what session it wants and can always request whatever session it wants in the getRealtimeEphemeralAPIKey. See the WebRTC example for how to do this.